pyproject.toml
[project]
name = "llm-eda"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llm-parallelization",
    "numpy>=2.2.6",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.2.0",
    "textattack>=0.3.10",
]

[tool.uv.sources]
llm-parallelization = {path = "/home/clyde/workspace/repositories/llm_parallelization", editable = true}

